requirements.txt
absl-py==0.8.1
altgraph==0.17
apipkg==1.5
appnope==0.1.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.0
boto==2.49.0
boto3==1.10.2
botocore==1.13.2
bs4==0.0.1
cachetools==3.1.1
certifi==2019.9.11
cffi==1.14.0
chainer==6.5.0
chardet==3.0.4
cleo==0.7.6
Click==7.0
clikit==0.4.2
cmudict==0.4.4
context2vec==1.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
decorator==4.4.0
defusedxml==0.6.0
demoji==0.1.5
Django==2.2.4
django-ssh-deployer==0.4.5
django-webpack-loader==0.6.0
djangorestframework==3.10.2
docutils==0.15.2
entrypoints==0.3
execnet==1.5.0
filelock==3.0.12
Flask==1.0.2
future==0.18.2
gast==0.2.2
gensim==3.8.1
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
idna==2.8
imageio==2.6.1
import-ipynb==0.1.3
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.2
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
language-check==1.1
macholib==1.14
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mlxtend==0.17.0
more-itertools==6.0.0
mysql-connector-python==8.0.16
mysqlclient==1.4.4
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
nltk==3.5
notebook==6.0.1
numpy==1.17.2
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.1
pastel==0.2.0
pefile==2019.4.18
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2018.11.26
pluggy==0.9.0
praw==6.5.1
prawcore==1.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycontractions==2.0.1
pycparser==2.19
pyemd==0.5.1
Pygments==2.4.2
PyInstaller==3.6
pylev==1.3.0
PyNaCl==1.3.0
pyparsing==2.4.2
PyPDF2==1.26.0
pyrsistent==0.15.4
pyspellchecker==0.5.4
pytest==4.3.1
pytest-cache==1.0
pytest-pep8==1.0.6
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.0
qtconsole==4.5.5
regex==2020.1.8
requests==2.23.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.1
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
smart-open==1.8.4
soupsieve==2.0
sqlparse==0.3.0
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.2
typing==3.6.6
typing-extensions==3.6.6
update-checker==0.16
urllib3==1.25.6
virtualenv==16.7.4
virtualenv-clone==0.5.3
vsearch==1.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.15.1
widgetsnbextension==3.5.1
word2vec==0.10.2
wrapt==1.11.2
XlsxWriter==1.2.8
